requirements.txt
Flask==3.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
